% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_demand}
\alias{food_demand}
\title{Food Demand Dataset}
\format{
A data.frame of size 1,427 x 4 containing
\itemize{
\item fridge_id
\item week_date
\item index
\item sold
}
}
\usage{
food_demand
}
\description{
Sales time series from smart fridges. The data was preprocessed and cleaned using forecast::tsclean.
}
\keyword{datasets}
