% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_avg.cv.R
\name{plot.ts_avg.cv}
\alias{plot.ts_avg.cv}
\title{plot.ts_avg.cv}
\usage{
\method{plot}{ts_avg.cv}(object, plot_type = c("cv", "best", "error"), type = NULL)
}
\arguments{
\item{object}{result of ts_avg.cv function}

\item{plot_type}{cv results, or best forecasts, or error}

\item{type}{select only certain avg. types for plotting}
}
\value{
ggplot2 object
}
\description{
Plot ts_avg.cv result
}
\examples{
source("inst/my_ets.R")

x <- replicate(5,{
  len <- rpois(1,20)
  cbind(seq(rpois(1,5),length.out=len),matrix(rnorm(2*len),ncol = 2))
  },simplify=FALSE)


ts_avg_res <- ts_avg(x,my_ets,h=2,k=2,type="simple_avg", xtest_idx = 20)
plot(ts_avg_res)

ts_avg_res_2 <- ts_avg(lapply(x,"[",TRUE,1:2),my_ets,h=2,k=1,type="simple_avg", xtest_idx = 15)
plot(ts_avg_res_2,h=1)

# cv examples
res_mv <- ts_avg.cv(x,my_ets,min_ts=10,k_grid=2:5,xtest_idx = 20,verbose=TRUE)

plot(res_mv, "cv")
plot(res_mv, "error")

res_uv <- ts_avg.cv(lapply(x,"[",TRUE,1:2),my_ets,min_ts=10,k_grid=2:5, xtest_idx=20)

plot(res_uv, "best")
plot(res_uv, "cv")
}
